move vnext script to azure-sdk-tools/gh_tools per feedback
