+++ eng/test_tools.txt
@@ -12,7 +12,6 @@ protobuf==3.17.3; python_version == '2.7'
 chardet>=2.0,<5.0
 cmarkgfm<0.7.0
 typing-extensions<=4.6.3
-PyGithub==1.59.0
 
 # locking packages defined as deps from azure-sdk-tools or azure-devtools
 pytoml==0.1.21
