+++ eng/tox/run_pylint.py
@@ -17,8 +17,7 @@ import sys
 from ci_tools.environment_exclusions import is_check_enabled
 from ci_tools.parsing import ParsedSetup
 from ci_tools.variables import in_ci
-
-from vnext_issue_creator import create_vnext_issue
+from gh_tools.vnext_issue_creator import create_vnext_issue
 
 logging.getLogger().setLevel(logging.INFO)
